+++ src/Model/Category.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 
 namespace WEM\AudioTracksBundle\Model;
 
+use Contao\Environment;
 use WEM\UtilsBundle\Model\Model;
 
 /**
@@ -27,4 +28,35 @@ class Category extends Model
      * @var string
      */
     protected static $strTable = 'tl_wem_audiotrack_category';
+
+    /**
+     * RSS Folder path
+     * 
+     * @var string
+     */
+    protected static $strRssFolder = 'share/audiotracks/rss/';
+
+    /**
+     * Generate URL for RSS
+     */
+    public function getRssFeedUrl(): string
+    {
+        if (!$this->rssFilename) {
+            return "Cannot generate RSS Feed url as category does not have a RSS filename setup";
+        }
+
+        return Environment::get('base') . static::$strRssFolder . $this->rssFilename;
+    }
+
+    /**
+     * Generate path for RSS
+     */
+    public function getRssFeedPath(): string
+    {
+        if (!$this->rssFilename) {
+            return "Cannot generate RSS Feed url as category does not have a RSS filename setup";
+        }
+
+        return 'web/' . static::$strRssFolder . $this->rssFilename;
+    }
 }