+++ src/Resources/contao/dca/tl_wem_audiotrack_tag.php
@@ -2,10 +2,12 @@
 
 declare(strict_types=1);
 
+use Contao\DC_Table;
+
 $GLOBALS['TL_DCA']['tl_wem_audiotrack_tag'] = [
     // Config
     'config' => [
-        'dataContainer' => 'Table',
+        'dataContainer' => DC_Table::class,
         'ptable' => 'tl_wem_audiotrack',
         'sql' => [
             'keys' => [